+++ user_docs/en/changes.md
@@ -6,9 +6,16 @@
 
 ### New Features
 
+* A new command, assigned to `NVDA+x`, has been introduced to repeat the last information spoken by NVDA; pressing it twice shows it in a browseable message. (#625, @CyrilleB79)
+* Added an unassigned command to toggle keyboard layout. (#19211, @CyrilleB79)
+
 ### Changes
 
+* It is now possible to open the log viewer with `NVDA+f1`, even when the log level is set to "disabled". (#19318, @CyrilleB79)
+
 ### Bug Fixes
+
+* Fixed excessive resource usage and highlight flickering when using Visual Highlight. (#17434, @hwf1324)
 
 ### Changes for Developers